\name{NullSens-package}
\alias{NullSens-package}
\docType{package}
\title{NullSens: Partitioning Abiotic and Biotic Contributions to Community Variation}
\description{NullSens is a method that employs a null model based on existing species covariation models to assess community-wide biotic covariation after removing individual species responses to environmental gradients. Analytical output from NullSens provides the percentages of the variation that can be explained due to abiotic, biotic and unexplained factors. Within the biotic variation results, the algorithm also outputs if it detects positive, negative, or mixed community covariation. NullSens can be used to verify hypotheses on species interactions when environmental gradients are present.}
\author{Steven D. Essinger <sde22@drexel.edu>, Christopher B. Blackwood, Gail L. Rosen}
\references{S.D. Essinger, "Partitioning Abiotic and Biotic Contributions to Community Variation," Ph.D. dissertation, Dept. Elect. Eng., Drexel Univ., Philadelphia, PA, 2013}